feat(chat): pod watch
